+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - [#412](https://github.com/influxdb/telegraf/pull/412): Additional memcached stats. Thanks @mgresser!
 - [#410](https://github.com/influxdb/telegraf/pull/410): Additional redis metrics. Thanks @vlaadbrain!
 - [#414](https://github.com/influxdb/telegraf/issues/414): Jolokia plugin auth parameters
+- [#415](https://github.com/influxdb/telegraf/issues/415): memcached plugin: support unix sockets
 
 ### Bugfixes
 - [#405](https://github.com/influxdb/telegraf/issues/405): Prometheus output cardinality issue