+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## v0.2.4 [unreleased]
+## v0.2.4 [2015-12-08]
 
 ### Features
 - [#412](https://github.com/influxdb/telegraf/pull/412): Additional memcached stats. Thanks @mgresser!